requirements.txt
azure-identity>=1.15.0
azure-mgmt-network>=25.2.0
azure-mgmt-resource>=23.1.0